v1.0.68

+++ includes/admin/class-import-export.php
@@ -341,8 +341,12 @@ class Import_Export implements Runner {
 
 		// Parse Options.
 		$wp_filesystem = WordPress::get_filesystem();
-		$settings      = $wp_filesystem->get_contents( $file['file'] );
-		$settings      = json_decode( $settings, true );
+		if ( is_null( $wp_filesystem ) ) {
+			return false;
+		}
+
+		$settings = $wp_filesystem->get_contents( $file['file'] );
+		$settings = json_decode( $settings, true );
 
 		\unlink( $file['file'] );
